backend/requirements.txt
asgiref==3.3.4
attrs==21.2.0
black==21.11b1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2
django-colorfield==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
pkg_resources==0.0.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
python3-openid==3.2.0
pytz==2021.1
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
